+++ SourceryTests/Generating/JavaScriptTemplateSpecs.swift
@@ -41,6 +41,17 @@ class JavaScriptTemplateTests: QuickSpec {
                 print("expected:\n\(expectedResult)\n\ngot:\n\(result)")
                 expect(result).to(equal(expectedResult))
             }
+            
+            it("provides typealias information") {
+                let templatePath = Stubs.jsTemplates + Path("Typealiases.ejs")
+                let expectedResult = try? (Stubs.resultDirectory + Path("Typealiases.swift")).read(.utf8)
+                
+                expect { try Sourcery(cacheDisabled: true).processFiles(.sources(Paths(include: [Stubs.sourceDirectory])), usingTemplates: Paths(include: [templatePath]), output: output, baseIndentation: 0) }.toNot(throwError())
+
+                let result = (try? (outputDir + Sourcery().generatedPath(for: templatePath)).read(.utf8))
+                print("expected:\n\(expectedResult)\n\ngot:\n\(result)")
+                expect(result).to(equal(expectedResult))
+            }
 
             it("handles includes") {
                 let templatePath = Stubs.jsTemplates + Path("Includes.ejs")