+++ SourceryTests/Stub/Result/Typealiases.swift
@@ -1,0 +1,7 @@
+// Generated using Sourcery Major.Minor.Patch — https://github.com/krzysztofzablocki/Sourcery
+// DO NOT EDIT
+// Typealiases in Bar
+// - name 'List', type '[Foo]'
+// Typealiases in Foo
+// - name 'Name', type 'String'
+// Typealiases in FooSubclass